CHORE: 기존 감정구슬 선택, 추천 루틴 화면를 emotion/component/template으로 이동
